% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ipcc_template.R
\name{build_ipcc_template}
\alias{build_ipcc_template}
\title{build_ipcc_template: convert the iTEM MIP3 database into the format for the IPCC AR6 inter-comparison}
\usage{
build_ipcc_template(
  item_data_fn,
  item_data = NA,
  ipcc_mapping_fn = "ipcc/item_to_ipcc_mapping.csv",
  output_folder
)
}
\arguments{
\item{item_data}{data table with iTEM template data}

\item{ipcc_mapping}{table with variable assignments from iTEM to IPCC AR6 template}
}
\description{
build_ipcc_template: convert the iTEM MIP3 database into the format for the IPCC AR6 inter-comparison
}
\details{
This function performs a post-hoc translation from the iTEM MIP3 output to
the IPCC AR6 template
}
