% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_MODEL.R
\name{correct_MODEL}
\alias{correct_MODEL}
\title{Apply corrections to MODEL submitted data}
\usage{
correct_MODEL(x)
}
\arguments{
\item{x}{data table with specific model data}
}
\description{
Apply corrections to MODEL submitted data
}
\details{
Applies model-specific corrections to reporting categories. This function should be renamed to the actual
  model name, and corrections applied.
}
