% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-downscale.R
\name{load_preprocessed_data}
\alias{load_preprocessed_data}
\title{Load the data submitted by the individual models}
\usage{
load_preprocessed_data(model_data_folder, model_names)
}
\arguments{
\item{model_data_folder}{folder where the pre-processed model output data is located}

\item{model_names}{names of the models (case sensitive)}
}
\description{
Load model-submitted data into a list of data frames where each element is a single model's output (i.e., all
scenarios in one data frame per model).
}
\details{
Requires that the model output data are generally cleaned and appropriately formatted, and placed into a
  common folder with sub-folders for each model. Individual data files are typically the model name plus the data
  submissions version number, and the extension ".csv". The function will pull the data with the highest version
  number.
}
